Use libunwind instead of glibc backtrace
glibc backtrace implementation has the following problems:

 - some addresses are not resolved to function names (so the backtrace
 - is sometimes incomplete);

 - signal-safety can be achieved but is not documented nor guarnateed;
   actually we use an undocumented hack for it;

 - to perform demangling, we have to parse implementation-specific
   text format of the backtrace output.

On the other hand, we now have a backtrace implementation using
libunwind, which does not have any of these problems.

This commit removes the glibc backtrace implementation and enables
libunwind dependency by default. This commit also adds a portable
wrapper for symbol demangling and employs it both in libunwind and
bionic implementations of backtrace.