Prohibit multiple ports usage in roc-recv
This PR aims to simplify roc-recv and roc-send usage.

Now, if we add some option to the command line interface we will not need
to think about to which port this option should be applied. Previously, we
allowed to use multiple source and repair ports and assumed that the
provided option would be applied to all ports.

Relates: #181, fe94ca3