Add guards to Semaphore, Mutex, and Cond
POSIX allows to destroy semaphore immediately after returning from
last wait() call, even if post() did not return yet.

This requirement assumes that post() does not use semaphore memory
after waking up the waiting thread. The implementation of this
requirement is possible, but is tricky enough.

Recent glibc and musl versions, and probably other platforms, fulfill
the requirement, but there are also platforms that don't support it,
notably macOS. On these envrionments, destroying a semaphore while
concurrent post() call is still running may lead to a crash or EINVAL.

The same is true for condition variables and mutexes.

To deal with it, we now implement an atomic guard and a spin-loop in
destructor, to wait until concurrent calls are fully finished.