Rework IResampler interface
Curcular buffer for input frames (to be resampled) is now an
implementation detail of resampler backend and is hidden by IResampler
interface.

As the result, we got rid of the two unnecessary frames for speex
resampler and decreased its latency accordingly. In addition, the
implementation of ResamplerReader and ResamplerWriter became simpler.