Fix race in ctl::TaskQueue
It was possible in some cases that TaskQueue will call handler twice
when cancelling a task. By the time of the second call the task may
be already destroyed.

To fix this race, we have switched from two atomics for task
state (state + in_ready_queue) to a single atomic (state).