Move all calls to wallclock from public_api to roc_pipeline
This commit implements two new pipeline options:
 - sender: enable_auto_cts: when enabled, SenderLoop automatically
   sets CTS of all frames to current time
 - receiver: enable_auto_reclock: when enabled, ReceiverLoop
   automatically invokes reclock() before returning frame, with
   current time

When SenderLoop and ReceiverLoop are used from CLI tools, these
options are disabled, because Pump manages CTS and reclock manually.

When SenderLoop and ReceiverLoop are used from C API, these options
are enabled, because public API has no notion of CTS and reclock.

With this approach, all work with wallclock time is isolated in
SenderLoop, ReceiverLoop, and Pump.