gh-86 Use script to generate downmixing tables
This commits adds a script "channel_tables_gen.py" that generates
all ChanMapTables tables (and some others).

The generated output is exactly the same as the existing tables.
(It will change in next commits).

Having this script allows to add more tables without lots of
manual work. We're going to add more tables in future commits.