gh-86 Cascade downmixing/upmixing
ChannelMapperMatrix now can build matrix from multiple tables
if a single table for direct mapping is not available.
It also can combine upmixing/downmixing if needed.

This allows to reduce number of tables significantly since
we don't need to cover all mask combinations now. This will
be especially important when we add more masks.

This commit also expands tests and fixes some coefficients
in downmixing tables, to ensure that cascade down/up-mixing
gives the same result and a single-step one.