[spo] gh-183 Report initialization errors via init_status()
This commit implements a part of gh-183 (pipeline error forwarding),
needed for gh-614 (status codes for frame writers/readers).

It replaces is_valid() in pipeline elements with init_status(), and
forwards initialization errors through pipeline.

Without it, when a function should forward a status, and both
initialization and read/write errors can happen, it wouldn't be
clear what status to return in case of initialization error.

Sponsored-by: waspd