Rollback from --min/--max-latency to --latency-tolerance
Earlier, in 0.3.0, we had --latency-tolerance.

Then we replaced it with --min-latency/--max-latency as part
of the preparations for adaptive latency. This change was not
part of any release yet.

Upcoming adaptive latency feature will change it yet another time:
    --latency-tolerance will be reintroduced again and used for
      bounds checking instead of min/max, as earlier
      (i.e. back to 0.3.0 behavior)
    --min-latency/--max-latency will be also kept, but used only
      for adaptive latency mode to define boundaries for target
      (new behavior specific to adaptive latency)

Since we're now releasing 0.4.0 without adaptive latency, this
commit temporarily removes --min-latency/--max-latency and returns
--latency-tolerance as it was in 0.3.0 and as it should be after
we add adaptive latency.

Otherwise we would break compatibility twice, in 0.4.0 and then
when adaptive latency is added.