gh-537: roc_pipeline benchmark: fix a task completer race
In peakload benchmarks it is possible for the TaskThread object to be
removed along with the pipeline_task_completed() callback before all
tasks are completed. This results in a SIGABRT with "pure virtual method
called" message.

To fix this move pipeline_task_completed() to TestPipeline class that has
a longer lifespan.

Fixes: #537